+++ src/gdsii.cpp
@@ -57,10 +57,10 @@ ErrorCode gdsii_read_record(FILE* in, uint8_t* buffer, uint64_t& buffer_count) {
     big_endian_swap16((uint16_t*)buffer, 1);  // second word is interpreted byte-wise (no swaping);
     const uint32_t record_length = *((uint16_t*)buffer);
     if (record_length < 4) {
-        DEBUG_PRINT("Record length should be at beast 4. Found %" PRIu32 "\n", record_length);
+        DEBUG_PRINT("Record length should be at least 4. Found %" PRIu32 "\n", record_length);
         fputs("[GDSTK] Invalid or corrupted GDSII file.\n", stderr);
         buffer_count = read_length;
-        return ErrorCode::InputFileError;
+        return ErrorCode::InvalidFile;
     } else if (record_length == 4) {
         buffer_count = read_length;
         return ErrorCode::NoError;