+++ src/gdsii.h
@@ -95,8 +95,9 @@ uint64_t gdsii_real_from_double(double value);
 
 double gdsii_real_to_double(uint64_t real);
 
-// Read record and make necessary swaps.  The size of the buffer must be passed
-// in buffer_count.  On return, the record lenght is returned in buffer_count.
+// Read a record and swaps only first 2 bytes (record length).  The size of the
+// buffer must be passed in buffer_count.  On return, the record lenght
+// (including header) is returned in buffer_count.
 ErrorCode gdsii_read_record(FILE* in, uint8_t* buffer, uint64_t& buffer_count);
 
 }  // namespace gdstk